requirements.txt
Django==5.2.12
djangorestframework==3.17.0
psycopg2-binary==2.9.9
celery==5.3.4
redis==5.0.1
openpyxl==3.1.2
django-celery-results==2.5.1